Moved DAOs to specific folders for json and postgres
