+++ CodingPractice/PythonAssignments/shoppingcart/dao/json/CustomerJsonDao.py
@@ -1,4 +1,4 @@
-from CodingPractice.PythonAssignments.shoppingcart.dao.JsonFileReader import JsonFileReader
+from CodingPractice.PythonAssignments.shoppingcart.dao.json.JsonFileReader import JsonFileReader
 from CodingPractice.PythonAssignments.shoppingcart.dao.CustomerDao import *
 from CodingPractice.PythonAssignments.shoppingcart.dao.CustomerDetailsDao import *
