+++ CodingPractice/PythonAssignments/shoppingcart/dao/json/OrderJsonDao.py
@@ -1,4 +1,4 @@
-from CodingPractice.PythonAssignments.shoppingcart.dao.JsonFileReader import JsonFileReader
+from CodingPractice.PythonAssignments.shoppingcart.dao.json.JsonFileReader import JsonFileReader
 from CodingPractice.PythonAssignments.shoppingcart.dao.OrderDao import OrderDao
 from CodingPractice.PythonAssignments.shoppingcart.domain.OrderDto import OrderDto, OrderLineDto
